backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# AI/ML
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5
scikit-learn==1.4.0
xgboost==2.0.3
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3
pandas==2.2.0

# HTTP & API
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# File Processing
python-magic==0.4.27
xmltodict==0.13.0

# Utilities
python-dotenv==1.0.0
pytz==2024.1
python-dateutil==2.8.2

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Message Queue
celery==5.3.6
kombu==5.3.5

# Task Scheduling
apscheduler==3.10.4

# S3/MinIO
boto3==1.34.34
minio==7.2.3

# Vector DB
pinecone-client==3.0.2
weaviate-client==4.4.0

# WebSocket
websockets==12.0

